Start with open folds also for (split) :VCSAnnotate!
The :setlocal foldlevel=999 gets cleared by the zR command; because of the former, this actually isn't necessary.
Also, the split blame header needs have the foldlevel set, too.
While we're at that, change the :set scrollbind to :setlocal (and then repeat in the split blame header), to avoid that this accidentally get inherited to windows split from there.