Turn off modeline processing on single-buffer annotate to avoid that the filetype gets changed
A modeline may set a filetype, but the extraction of the predecessor version (by the different VCS types) checks for the {TYPE}annotate filetype, so it must not be changed for further drill-down to work.
The syntax highlighting that is enabled by a ft= / syntax= modeline usually is broken by the prepended annotations, anyway. As this buffer is only for browsing, not modifications, any other filetype settings do not make much sense, too.
Therefore, suppress modelines for single-buffer annotations. As the annotate action opens a new (scratch) buffer, we have to globally disable modelines; just a :setlocal nomodeline wouldn't work.